move ReviewPage from page to component folder
