+++ src/components/ReviewPage/components/AllReviews/components/ToolBar/components/BarButton/BarButton.tsx
@@ -24,6 +24,8 @@ const StyledButton = styled.button`
   font-size: 18px;
   font-weight: 400;
 
+  cursor: pointer;
+
   &:hover {
     background-color: ${({ theme }) => theme.color.main};
     color: #3d3d3d; // TODO: manage the color using theme
